Updates to clean up javadocs of the SDK
