+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/BadRequestException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class BadRequestException extends BridgeServerException {
+public class BadRequestException extends BridgeSDKException {
 
     public BadRequestException(String message, String endpoint) {
         super(message, 400, endpoint);
