+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/ConcurrentModificationException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class ConcurrentModificationException extends BridgeServerException {
+public class ConcurrentModificationException extends BridgeSDKException {
 
     public ConcurrentModificationException(String message, String endpoint) {
         super(message, 409, endpoint);
