+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/EntityAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class EntityAlreadyExistsException extends BridgeServerException {
+public class EntityAlreadyExistsException extends BridgeSDKException {
 
     public EntityAlreadyExistsException(String message, String endpoint) {
         super(message, 409, endpoint);
