+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/EntityNotFoundException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class EntityNotFoundException extends BridgeServerException {
+public class EntityNotFoundException extends BridgeSDKException {
 
     public EntityNotFoundException(String message, String endpoint) {
         super(message, 404, endpoint);
