+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/NotAuthenticatedException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class NotAuthenticatedException extends BridgeServerException {
+public class NotAuthenticatedException extends BridgeSDKException {
 
     public NotAuthenticatedException(String message, String endpoint) {
         super(message, 401, endpoint);
