+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/PublishedSurveyException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class PublishedSurveyException extends BridgeServerException {
+public class PublishedSurveyException extends BridgeSDKException {
 
     public PublishedSurveyException(String message, String url) {
         super(message, 400, url);
