+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/UnauthorizedException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class UnauthorizedException extends BridgeServerException {
+public class UnauthorizedException extends BridgeSDKException {
 
     public UnauthorizedException(String message, String endpoint) {
         super(message, 403, endpoint);
