+++ src/main/java/org/sagebionetworks/bridge/sdk/exceptions/UnsupportedVersionException.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.bridge.sdk.exceptions;
 
 @SuppressWarnings("serial")
-public class UnsupportedVersionException extends BridgeServerException {
+public class UnsupportedVersionException extends BridgeSDKException {
     
     public UnsupportedVersionException(String message, String endpoint) {
         super(message, 410, endpoint);
