+++ src/main/java/org/sagebionetworks/bridge/sdk/utils/Utilities.java
@@ -1,9 +1,8 @@
-package org.sagebionetworks.bridge.sdk;
+package org.sagebionetworks.bridge.sdk.utils;
 
 import static org.apache.commons.validator.routines.UrlValidator.ALLOW_LOCAL_URLS;
 import static org.apache.commons.validator.routines.UrlValidator.NO_FRAGMENTS;
 
-import java.io.IOException;
 import java.util.Map;
 
 import com.fasterxml.jackson.core.type.TypeReference;
@@ -11,8 +10,6 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.commons.lang3.builder.ToStringStyle;
 import org.apache.commons.validator.routines.EmailValidator;
 import org.apache.commons.validator.routines.UrlValidator;
-import org.apache.http.client.fluent.Request;
-import org.apache.http.client.fluent.Response;
 import org.sagebionetworks.bridge.sdk.json.LowercaseEnumModule;
 import org.sagebionetworks.bridge.sdk.models.holders.GuidCreatedOnVersionHolder;
 import org.sagebionetworks.bridge.sdk.models.holders.GuidHolder;
@@ -71,34 +68,17 @@ public final class Utilities {
         return MAPPER;
     }
 
-    static boolean isValidEmail(String email) {
+    public static boolean isValidEmail(String email) {
         if (email == null) {
             throw new IllegalArgumentException("Email cannot be null.");
         }
         return emailValidator.isValid(email);
     }
 
-    static boolean isValidUrl(String url) {
+    public static boolean isValidUrl(String url) {
         if (url == null) {
             throw new IllegalArgumentException("URL cannot be null.");
         }
         return urlValidator.isValid(url);
-    }
-
-    static boolean isConnectableUrl(String url, int timeout) {
-        if (!isValidUrl(url)) {
-            throw new IllegalArgumentException("URL is not a valid one: " + url);
-        } else if (timeout <= 0 || 10 * 1000 <= timeout) {
-            throw new IndexOutOfBoundsException("timeout isn't in the valid range (0 < timeout < 10 minutes): "
-                    + timeout);
-        }
-        url = url.replaceFirst("https", "http");
-        try {
-            Response response = Request.Head(url).connectTimeout(timeout).execute();
-            int statusCode = response.returnResponse().getStatusLine().getStatusCode();
-            return (200 <= statusCode && statusCode <= 399);
-        } catch (IOException exception) {
-            return false;
-        }
     }
 }
