+++ src/test/java/org/sagebionetworks/bridge/sdk/models/schedules/ActivityTest.java
@@ -7,7 +7,8 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
-import org.sagebionetworks.bridge.sdk.Utilities;
+
+import org.sagebionetworks.bridge.sdk.utils.Utilities;
 
 public class ActivityTest {
