+++ src/test/java/org/sagebionetworks/bridge/sdk/models/studies/EmailTemplateTest.java
@@ -4,8 +4,9 @@ import static org.junit.Assert.assertEquals;
 import nl.jqno.equalsverifier.EqualsVerifier;
 
 import org.junit.Test;
-import org.sagebionetworks.bridge.sdk.Utilities;
+
 import org.sagebionetworks.bridge.sdk.models.studies.EmailTemplate.MimeType;
+import org.sagebionetworks.bridge.sdk.utils.Utilities;
 
 import com.fasterxml.jackson.databind.JsonNode;
