+++ src/test/java/org/sagebionetworks/bridge/sdk/models/studies/StudyTest.java
@@ -5,8 +5,9 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 
 import org.junit.Test;
-import org.sagebionetworks.bridge.sdk.Utilities;
+
 import org.sagebionetworks.bridge.sdk.models.studies.EmailTemplate.MimeType;
+import org.sagebionetworks.bridge.sdk.utils.Utilities;
 
 import com.fasterxml.jackson.databind.JsonNode;
