+++ src/test/java/org/sagebionetworks/bridge/sdk/models/upload/UploadFieldDefinitionTest.java
@@ -9,8 +9,8 @@ import java.util.Map;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.Test;
 
-import org.sagebionetworks.bridge.sdk.Utilities;
 import org.sagebionetworks.bridge.sdk.exceptions.InvalidEntityException;
+import org.sagebionetworks.bridge.sdk.utils.Utilities;
 
 public class UploadFieldDefinitionTest {
     @Test(expected = InvalidEntityException.class)
